feat(engineer-selector): only show remove icon on hover

+++ frontend/components/EngineerAssignment.vue
@@ -9,16 +9,19 @@ const props = defineProps<{
     current: Engineer | undefined;
 }>();
 
+const previousValue = ref<string>(props.current?.id ?? '');
+
 const onChange = async (changeEvent: SelectChangeEvent) => {
+    if (changeEvent.value == previousValue.value) return;
+
     if (changeEvent.value == null) {
         await engineerStore.bulkDelete([props.activityKey], props.assignAs);
-        return;
-    }
-    if (props.assignAs === "delegator") {
+    } else if (props.assignAs === "delegator") {
         await engineerStore.setDelegatorOnActivities([props.activityKey], changeEvent.value);
     } else if (props.assignAs === "engineer") {
         await engineerStore.setEngineerOnActivities([props.activityKey], changeEvent.value);
     }
+    previousValue.value = changeEvent.value;
 }
 
 onMounted(() => {
@@ -28,6 +31,7 @@ onMounted(() => {
 
 <template>
     <Select
+        class="engineer-assignment"
         :options="assignAs === 'delegator' ? engineerStore.delegators : engineerStore.engineers"
         option-value="id"
         option-label="name"
@@ -37,9 +41,31 @@ onMounted(() => {
         filter
         show-clear
         @change="onChange"
+        style="width: 200px"
+        size="small"
     >
 
     </Select>
 </template>
 
-<style lang="scss" scoped></style>
+<style lang="scss">
+.p-select.engineer-assignment {
+    .p-select-clear-icon {
+        display: none;
+    }
+
+    .p-select-label {
+        padding-inline-end: 0;
+    }
+}
+
+.p-select.engineer-assignment:hover {
+    .p-select-clear-icon {
+        display: block;
+    }
+
+    .p-select-label {
+        padding-inline-end: 28px;
+    }
+}
+</style>
